pyproject.toml
[tool.poetry]
name = "qaguru-python-4-api-tests"
version = "0.1.0"
description = ""
authors = ["Artem Yurchenko <yourchencore@yandex.ru>"]
readme = "README.md"
packages = [{ include = "qaguru_python_4_api_tests" }]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
pytest = "^7.3.1"
voluptuous = "^0.13.1"
pytest-voluptuous = "^1.2.0"
allure-pytest = "^2.13.1"
curlify = "^2.2.1"
python-dotenv = "0.20.0"
selene = "2.0.0rc2"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = 1
log_cli_format = "%(asctime)s %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_level = "INFO"
addopts = " -s --clean-alluredir --alluredir=allure-results --allure-no-capture "